Update 1.3 -> 1.5
Capes, bug fixed to guis, finally added setting for flower weapons blocking, blocking fishingTimer and autofish on carnival area and many more

+++ src/main/java/treemek/mesky/features/illegal/GhostPickaxe.java
@@ -35,8 +35,8 @@ public class GhostPickaxe {
             for (int i = 0; i < inventory.mainInventory.length; i++) {
                 ItemStack item = inventory.mainInventory[i];
                 if (item != null && item.getItem() == Items.golden_pickaxe && i != SettingsConfig.GhostPickaxeSlot.number) {
-                	if(SettingsConfig.GhostPickaxeSlot.number < 1) SettingsConfig.GhostPickaxeSlot.number = 1;
-                    inventory.setInventorySlotContents(SettingsConfig.GhostPickaxeSlot.number-1, item.copy());
+                	if(SettingsConfig.GhostPickaxeSlot.number < 1) SettingsConfig.GhostPickaxeSlot.number = 1D;
+                    inventory.setInventorySlotContents((int)(SettingsConfig.GhostPickaxeSlot.number-1), item.copy());
                     break; // Exit loop after transferring the pickaxe
                 }
             }
